+++ system/modules/core/dca/tl_log.php
@@ -180,6 +180,17 @@ class tl_log extends Backend
 			case 'ERROR':
 				$label = preg_replace('@^(.*</span> )(.*)$@U', '$1 <span class="tl_red">$2</span>', $label);
 				break;
+
+			default:
+				if (isset($GLOBALS['TL_HOOKS']['colorizeLogEntries']) && is_array($GLOBALS['TL_HOOKS']['colorizeLogEntries']))
+				{
+					foreach ($GLOBALS['TL_HOOKS']['colorizeLogEntries'] as $callback)
+					{
+						$this->import($callback[0]);
+						$label = $this->$callback[0]->$callback[1]($row, $label);
+					}
+				}
+				break;
 		}
 
 		return $label;